Separate Java-related logic from common libraries
- Split resource connector manager library
- Split database library
- Remove unused dependency from java-related modules for port of libraries
