+++ PluginsAndFeatures/azure-toolkit-for-intellij/.idea/gradle.xml
@@ -19,6 +19,7 @@
             <option value="$PROJECT_DIR$/azure-intellij-plugin-containerservice" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-cosmos" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-database" />
+            <option value="$PROJECT_DIR$/azure-intellij-plugin-database-java" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-guidance" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-lib" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-lib-java" />
@@ -29,6 +30,7 @@
             <option value="$PROJECT_DIR$/azure-intellij-plugin-vm" />
             <option value="$PROJECT_DIR$/azure-intellij-resource-connector-aad" />
             <option value="$PROJECT_DIR$/azure-intellij-resource-connector-lib" />
+            <option value="$PROJECT_DIR$/azure-intellij-resource-connector-lib-java" />
             <option value="$PROJECT_DIR$/azure-sdk-reference-book" />
             <option value="$PROJECT_DIR$/buildSrc" />
           </set>
@@ -42,12 +44,14 @@
           <set>
             <option value="$PROJECT_DIR$/azure-intellij-plugin-base" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-database" />
+            <option value="$PROJECT_DIR$/azure-intellij-plugin-database-java" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-lib" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-lib-java" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-redis" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-service-explorer" />
             <option value="$PROJECT_DIR$/azure-intellij-plugin-storage" />
             <option value="$PROJECT_DIR$/azure-intellij-resource-connector-lib" />
+            <option value="$PROJECT_DIR$/azure-intellij-resource-connector-lib-java" />
           </set>
         </option>
       </GradleProjectSettings>
