+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-appservice/build.gradle
@@ -7,7 +7,6 @@ intellij {
 
 dependencies {
     compile project(':azure-intellij-plugin-lib')
-    compile project(':azure-intellij-plugin-lib-java')
     compile project(':azure-intellij-plugin-guidance')
     compile project(':azure-intellij-plugin-containerregistry')
     compile 'com.spotify:docker-client:8.16.0'
