+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-base/gradle.properties
@@ -15,6 +15,8 @@ org.gradle.daemon=true
 IntellijPluginBase=..
 IntellijPluginModules=\
   azure-intellij-resource-connector-lib,\
+  azure-intellij-resource-connector-lib-java,\
   azure-intellij-plugin-database,\
+  azure-intellij-plugin-database-java,\
   azure-intellij-plugin-storage,\
   azure-intellij-plugin-redis
