+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-containerservice/build.gradle
@@ -9,7 +9,7 @@ repositories {
 dependencies {
     compile project(':azure-intellij-plugin-lib')
     compile project(':azure-intellij-plugin-lib-java')
-    compile project(':azure-intellij-resource-connector-lib')
+    compile project(':azure-intellij-resource-connector-lib-java')
     compile 'com.spotify:docker-client:8.16.0'
     compile 'io.kubernetes:client-java:16.0.0'
     compile 'com.microsoft.azure:azure-toolkit-storage-lib'
