+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database-java/build.gradle
@@ -1,0 +1,7 @@
+intellij {
+}
+
+dependencies {
+    compile project(':azure-intellij-plugin-database')
+    compile project(':azure-intellij-resource-connector-lib-java')
+}
