+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-redis/build.gradle
@@ -3,7 +3,7 @@ intellij {
 
 dependencies {
     compile project(':azure-intellij-plugin-lib')
-    compile project(':azure-intellij-resource-connector-lib')
+    compile project(':azure-intellij-resource-connector-lib-java')
     compile 'com.microsoft.azure:azure-toolkit-redis-lib'
     compile 'com.microsoft.azure:azure-toolkit-ide-common-lib'
     compile 'com.microsoft.azure:azure-toolkit-ide-redis-lib:' + azureToolkitVersion
