+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage/build.gradle
@@ -3,7 +3,7 @@ intellij {
 
 dependencies {
     compile project(':azure-intellij-plugin-lib')
-    compile project(':azure-intellij-resource-connector-lib')
+    compile project(':azure-intellij-resource-connector-lib-java')
     compile 'com.microsoft.azure:azure-toolkit-storage-lib'
     compile 'com.microsoft.azure:azure-toolkit-ide-common-lib'
     compile 'com.microsoft.azure:azure-toolkit-ide-storage-lib:' + azureToolkitVersion
