+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/build.gradle
@@ -1,0 +1,8 @@
+intellij {
+    plugins = ['java', "properties"]
+}
+
+dependencies {
+    compile project(':azure-intellij-plugin-lib-java')
+    compile project(':azure-intellij-resource-connector-lib')
+}
