+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/build.gradle
@@ -1,9 +1,8 @@
 intellij {
-    plugins = ['java', "properties"]
+    plugins = [ 'properties' ]
 }
 
 dependencies {
     compile project(':azure-intellij-plugin-lib')
-    compile project(':azure-intellij-plugin-lib-java')
     compile 'com.microsoft.azure:azure-toolkit-ide-common-lib'
 }
