+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/src/main/java/com/microsoft/azure/toolkit/intellij/connector/Connection.java
@@ -5,29 +5,18 @@
 
 package com.microsoft.azure.toolkit.intellij.connector;
 
-import com.intellij.execution.application.ApplicationConfiguration;
-import com.intellij.execution.configurations.JavaParameters;
-import com.intellij.execution.configurations.ModuleBasedConfiguration;
 import com.intellij.execution.configurations.RunConfiguration;
 import com.intellij.openapi.actionSystem.DataContext;
-import com.intellij.openapi.module.Module;
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.intellij.common.runconfig.IWebAppRunConfiguration;
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
-import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import org.apache.commons.lang3.StringUtils;
 import org.jdom.Element;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
@@ -43,38 +32,35 @@ import java.util.stream.Collectors;
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 public class Connection<R, C> {
     public static final String ENV_PREFIX = "%ENV_PREFIX%";
-    private static final String SPRING_BOOT_CONFIGURATION = "com.intellij.spring.boot.run.SpringBootApplicationRunConfiguration";
+
     @Nonnull
     @EqualsAndHashCode.Include
     protected final Resource<R> resource;
+
     @Nonnull
     @EqualsAndHashCode.Include
     protected final Resource<C> consumer;
+
     @Nonnull
     @EqualsAndHashCode.Include
     protected final ConnectionDefinition<R, C> definition;
+
     @Setter
     @Getter(AccessLevel.NONE)
     private String envPrefix;
+
     private Map<String, String> env = new HashMap<>();
 
     /**
      * is this connection applicable for the specified {@code configuration}.<br>
      * - the {@code Connect Azure Resource} before run task will take effect if
-     * applicable: the {@link #prepareBeforeRun} & {@link #updateJavaParametersAtRun}
+     * applicable: the {@link #prepareBeforeRun}
      * will be called.
      *
      * @return true if this connection should intervene the specified {@code configuration}.
      */
     public boolean isApplicableFor(@Nonnull RunConfiguration configuration) {
-        final boolean javaAppRunConfiguration = configuration instanceof ApplicationConfiguration;
-        final boolean springbootAppRunConfiguration = StringUtils.equals(configuration.getClass().getName(), SPRING_BOOT_CONFIGURATION);
-        final boolean azureWebAppRunConfiguration = configuration instanceof IWebAppRunConfiguration;
-        if (javaAppRunConfiguration || azureWebAppRunConfiguration || springbootAppRunConfiguration) {
-            final Module module = getTargetModule(configuration);
-            return Objects.nonNull(module) && Objects.equals(module.getName(), this.consumer.getName());
-        }
-        return false;
+        return configuration instanceof IWebAppRunConfiguration;
     }
 
     public Map<String, String> getEnvironmentVariables(final Project project) {
@@ -100,29 +86,12 @@ public class Connection<R, C> {
         }
     }
 
-    /**
-     * update java parameters exactly before start the {@code configuration}
-     */
-    public void updateJavaParametersAtRun(@Nonnull RunConfiguration configuration, @Nonnull JavaParameters parameters) {
-        if (Objects.nonNull(this.env)) {
-            for (final Map.Entry<String, String> entry : this.env.entrySet()) {
-                parameters.addEnv(entry.getKey(), entry.getValue());
-            }
-        }
-        if (this.resource.getDefinition() instanceof IJavaAgentSupported) {
-            parameters.getVMParametersList()
-                    .add(String.format("-javaagent:%s", ((IJavaAgentSupported) this.resource.getDefinition()).getJavaAgent().getAbsolutePath()));
-        }
+    public boolean isEnvironmentSet() {
+        return Objects.nonNull(this.env);
     }
 
-    @Nullable
-    private static Module getTargetModule(@Nonnull RunConfiguration configuration) {
-        if (configuration instanceof ModuleBasedConfiguration) {
-            return ((ModuleBasedConfiguration<?, ?>) configuration).getConfigurationModule().getModule();
-        } else if (configuration instanceof IWebAppRunConfiguration) {
-            return ((IWebAppRunConfiguration) configuration).getModule();
-        }
-        return null;
+    public Set<Map.Entry<String, String>> getEnvironmentEntries() {
+        return env.entrySet();
     }
 
     public String getEnvPrefix() {
