+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/src/main/java/com/microsoft/azure/toolkit/intellij/connector/ConnectionRunnerForRunConfiguration.java
@@ -7,37 +7,23 @@ package com.microsoft.azure.toolkit.intellij.connector;
 
 import com.intellij.execution.BeforeRunTask;
 import com.intellij.execution.BeforeRunTaskProvider;
-import com.intellij.execution.RunConfigurationExtension;
-import com.intellij.execution.RunManagerEx;
-import com.intellij.execution.RunManagerListener;
-import com.intellij.execution.RunnerAndConfigurationSettings;
-import com.intellij.execution.configurations.JavaParameters;
 import com.intellij.execution.configurations.RunConfiguration;
-import com.intellij.execution.configurations.RunConfigurationBase;
-import com.intellij.execution.configurations.RunnerSettings;
-import com.intellij.execution.impl.ConfigurationSettingsEditorWrapper;
 import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.openapi.actionSystem.DataContext;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Key;
 import com.microsoft.azure.toolkit.ide.common.icon.AzureIcons;
 import com.microsoft.azure.toolkit.intellij.common.IntelliJAzureIcons;
-import com.microsoft.azure.toolkit.intellij.common.runconfig.IWebAppRunConfiguration;
 import com.microsoft.azure.toolkit.lib.common.messager.ExceptionNotification;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
-import com.microsoft.intellij.util.BuildArtifactBeforeRunTaskUtils;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.java.Log;
 import org.apache.commons.collections4.CollectionUtils;
-import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.swing.*;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -45,14 +31,21 @@ import java.util.stream.Collectors;
  */
 @Log
 public class ConnectionRunnerForRunConfiguration extends BeforeRunTaskProvider<ConnectionRunnerForRunConfiguration.MyBeforeRunTask> {
+
     private static final String NAME = "Connect Azure Resource";
+
     private static final String DESCRIPTION = "Connect Azure Resource";
+
     private static final Icon ICON = IntelliJAzureIcons.getIcon(AzureIcons.Common.AZURE);
+
     private static final Key<MyBeforeRunTask> ID = Key.create("ConnectionRunnerForConfigurationId");
+
     @Getter
     public String name = NAME;
+
     @Getter
     public Key<MyBeforeRunTask> id = ID;
+
     @Getter
     public Icon icon = ICON;
 
@@ -107,78 +100,6 @@ public class ConnectionRunnerForRunConfiguration extends BeforeRunTaskProvider<C
         public List<Connection<?, ?>> getConnections() {
             final List<Connection<?, ?>> connections = this.config.getProject().getService(ConnectionManager.class).getConnections();
             return connections.stream().filter(c -> c.isApplicableFor(config)).collect(Collectors.toList());
-        }
-    }
-
-    public static class MyRunConfigurationExtension extends RunConfigurationExtension {
-
-        @Override
-        @ExceptionNotification
-        @AzureOperation(name = "connector.setup_connection_for_configuration", type = AzureOperation.Type.ACTION)
-        public <T extends RunConfigurationBase<?>> void updateJavaParameters(@Nonnull T config, @Nonnull JavaParameters params, RunnerSettings s) {
-            config.getBeforeRunTasks().stream().filter(t -> t instanceof MyBeforeRunTask).map(t -> (MyBeforeRunTask) t)
-                .flatMap(t -> t.getConnections().stream())
-                .forEach(c -> c.updateJavaParametersAtRun(config, params));
-        }
-
-        @Override
-        public boolean isApplicableFor(@Nonnull RunConfigurationBase<?> configuration) {
-            return configuration.getBeforeRunTasks().stream().anyMatch(c -> c instanceof MyBeforeRunTask);
-        }
-    }
-
-    public static class BeforeRunTaskAdder implements RunManagerListener, ConnectionTopics.ConnectionChanged, IWebAppRunConfiguration.ModuleChangedListener {
-        @Override
-        @ExceptionNotification
-        public void runConfigurationAdded(@Nonnull RunnerAndConfigurationSettings settings) {
-            final RunConfiguration config = settings.getConfiguration();
-            final List<Connection<?, ?>> connections = config.getProject().getService(ConnectionManager.class).getConnections();
-            final List<BeforeRunTask<?>> tasks = config.getBeforeRunTasks();
-            if (connections.stream().anyMatch(c -> c.isApplicableFor(config)) && tasks.stream().noneMatch(t -> t instanceof MyBeforeRunTask)) {
-                config.getBeforeRunTasks().add(new MyBeforeRunTask(config));
-            }
-        }
-
-        @Override
-        public void runConfigurationChanged(@NotNull RunnerAndConfigurationSettings settings) {
-            this.artifactMayChanged(settings.getConfiguration(), null);
-        }
-
-        @Override
-        @ExceptionNotification
-        @AzureOperation(name = "connector.update_connection_task", type = AzureOperation.Type.ACTION)
-        public void connectionChanged(Project project, Connection<?, ?> connection, ConnectionTopics.Action change) {
-            final RunManagerEx rm = RunManagerEx.getInstanceEx(project);
-            final List<RunConfiguration> configurations = rm.getAllConfigurationsList();
-            for (final RunConfiguration config : configurations) {
-                final List<BeforeRunTask<?>> tasks = config.getBeforeRunTasks();
-                if (change == ConnectionTopics.Action.ADD) {
-                    if (connection.isApplicableFor(config) && tasks.stream().noneMatch(t -> t instanceof MyBeforeRunTask)) {
-                        tasks.add(new MyBeforeRunTask(config));
-                    }
-                } else {
-                    final List<Connection<?, ?>> connections = config.getProject().getService(ConnectionManager.class).getConnections();
-                    if (connections.stream().noneMatch(c -> c.isApplicableFor(config))) {
-                        tasks.removeIf(t -> t instanceof MyBeforeRunTask);
-                    }
-                }
-            }
-        }
-
-        @Override
-        @ExceptionNotification
-        public void artifactMayChanged(@Nonnull RunConfiguration config, @Nullable ConfigurationSettingsEditorWrapper editor) {
-            final List<Connection<?, ?>> connections = config.getProject().getService(ConnectionManager.class).getConnections();
-            final List<BeforeRunTask<?>> tasks = config.getBeforeRunTasks();
-            Optional.ofNullable(editor).ifPresent(e -> BuildArtifactBeforeRunTaskUtils.removeTasks(e, (t) -> t instanceof MyBeforeRunTask));
-            tasks.removeIf(t -> t instanceof MyBeforeRunTask);
-            if (connections.stream().anyMatch(c -> c.isApplicableFor(config))) {
-                final List<BeforeRunTask> newTasks = new ArrayList<>(tasks);
-                final MyBeforeRunTask task = new MyBeforeRunTask(config);
-                newTasks.add(task);
-                RunManagerEx.getInstanceEx(config.getProject()).setBeforeRunTasks(config, newTasks);
-                Optional.ofNullable(editor).ifPresent(e -> e.addBeforeLaunchStep(task));
-            }
         }
     }
 }
