+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/src/main/resources/META-INF/azure-intellij-resource-connector-lib.xml
@@ -17,17 +17,8 @@
             serviceInterface="com.microsoft.azure.toolkit.intellij.connector.ResourceManager"
             serviceImplementation="com.microsoft.azure.toolkit.intellij.connector.ResourceManager$Impl" />
         <stepsBeforeRunProvider implementation="com.microsoft.azure.toolkit.intellij.connector.ConnectionRunnerForRunConfiguration" id="ConnectionRunnerForConfigurationId"/>
-        <runConfigurationExtension implementation="com.microsoft.azure.toolkit.intellij.connector.ConnectionRunnerForRunConfiguration$MyRunConfigurationExtension"/>
-        <completion.contributor language="Properties" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.SpringPropertiesCompletionContributor"/>
-        <codeInsight.lineMarkerProvider language="Properties" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.SpringPropertiesLineMarkerProvider"/>
     </extensions>
     <projectListeners>
-        <listener class="com.microsoft.azure.toolkit.intellij.connector.ConnectionRunnerForRunConfiguration$BeforeRunTaskAdder"
-                  topic="com.intellij.execution.RunManagerListener" />
-        <listener class="com.microsoft.azure.toolkit.intellij.connector.ConnectionRunnerForRunConfiguration$BeforeRunTaskAdder"
-                  topic="com.microsoft.azure.toolkit.intellij.connector.ConnectionTopics$ConnectionChanged" />
-        <listener class="com.microsoft.azure.toolkit.intellij.connector.ConnectionRunnerForRunConfiguration$BeforeRunTaskAdder"
-                  topic="com.microsoft.azure.toolkit.intellij.common.runconfig.IWebAppRunConfiguration$ModuleChangedListener" />
         <listener class="com.microsoft.azure.toolkit.intellij.connector.ResourceConnectionExplorer$ToolWindowOpener"
                   topic="com.microsoft.azure.toolkit.intellij.connector.ConnectionTopics$ConnectionChanged" />
     </projectListeners>
