+++ PluginsAndFeatures/azure-toolkit-for-intellij/build.gradle
@@ -173,6 +173,7 @@ dependencies {
     compile project(':azure-intellij-plugin-lib-java')
     compile project(':azure-sdk-reference-book')
     compile project(':azure-intellij-resource-connector-lib')
+    compile project(':azure-intellij-resource-connector-lib-java')
     compile project(':azure-intellij-resource-connector-aad')
     compile project(':azure-intellij-plugin-vm')
     compile project(':azure-intellij-plugin-springcloud')
@@ -180,6 +181,7 @@ dependencies {
     compile project(':azure-intellij-plugin-redis')
     compile project(':azure-intellij-plugin-service-explorer')
     compile project(':azure-intellij-plugin-database')
+    compile project(':azure-intellij-plugin-database-java')
     compile project(':azure-intellij-plugin-appservice')
     compile project(':azure-intellij-plugin-arm')
     compile project(':azure-intellij-plugin-containerregistry')
