+++ PluginsAndFeatures/azure-toolkit-for-intellij/settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'azure-toolkit-for-intellij'
 include 'azure-intellij-plugin-lib'
 include 'azure-intellij-plugin-lib-java'
 include 'azure-intellij-resource-connector-lib'
+include 'azure-intellij-resource-connector-lib-java'
 include 'azure-sdk-reference-book'
 include 'azure-intellij-resource-connector-aad'
 include 'azure-intellij-plugin-springcloud'
@@ -10,6 +11,7 @@ include 'azure-intellij-plugin-vm'
 include 'azure-intellij-plugin-storage'
 include 'azure-intellij-plugin-redis'
 include 'azure-intellij-plugin-database'
+include 'azure-intellij-plugin-database-java'
 include 'azure-intellij-plugin-appservice'
 include 'azure-intellij-plugin-arm'
 include 'azure-intellij-plugin-containerregistry'
@@ -17,4 +19,3 @@ include 'azure-intellij-plugin-applicationinsights'
 include 'azure-intellij-plugin-guidance'
 include 'azure-intellij-plugin-containerservice'
 include 'azure-intellij-plugin-cosmos'
-
